+++ src/Controller/CompanyController.php
@@ -124,6 +124,7 @@ final class CompanyController extends AbstractController
         $this->denyAccessUnlessVerified($this->getUser());
 
         $company = new Company();
+        $company->setCreator($this->getUser());
         $form = $this->createForm(CompanyType::class, $company);
         $form->handleRequest($request);
 
@@ -192,10 +193,14 @@ final class CompanyController extends AbstractController
 
 
     #[Route('/company/{id}/edit', name: 'app_company_edit', methods: ['GET', 'POST'])]
-    // #[IsGranted('ROLE_USER')]
+    #[IsGranted('ROLE_USER')]
     public function edit(Request $request, Company $company, EntityManagerInterface $entityManager): Response
     {
         $this->denyAccessUnlessVerified($this->getUser());
+
+        if (!$this->isGranted('edit', $company)) {
+            throw new AccessDeniedException('Accès refusé.');
+        }
 
         $form = $this->createForm(CompanyType::class, $company);
         $form->handleRequest($request);
@@ -217,6 +222,10 @@ final class CompanyController extends AbstractController
     public function delete(Request $request, Company $company, EntityManagerInterface $entityManager): Response
     {
         $this->denyAccessUnlessVerified($this->getUser());
+
+        if (!$this->isGranted('edit', $company)) {
+            throw new AccessDeniedException('Accès refusé.');
+        }
 
         if ($this->isCsrfTokenValid('delete' . $company->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($company);
