+++ src/Controller/InternshipController.php
@@ -11,6 +11,7 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 #[Route('/internship')]
 final class InternshipController extends AbstractController
@@ -50,10 +51,10 @@ final class InternshipController extends AbstractController
 
     #[Route('/new', name: 'app_internship_new', methods: ['GET', 'POST'])]
     #[IsGranted('ROLE_USER')]
-
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $internship = new Internship();
+        $internship->setCreator($this->getUser());
         $form = $this->createForm(InternshipType::class, $internship);
         $form->handleRequest($request);
 
@@ -74,7 +75,6 @@ final class InternshipController extends AbstractController
 
     #[Route('/pending', name: 'internship_pending')]
     #[IsGranted('ROLE_TEACHER')]
-
     public function pending(EntityManagerInterface $entityManager): Response
     {
         $internships = $entityManager->getRepository(Internship::class)->findBy(['is_verified' => false]);
@@ -86,7 +86,6 @@ final class InternshipController extends AbstractController
 
     #[Route('/verify/{id}', name: 'internship_verify')]
     #[IsGranted('ROLE_TEACHER')]
-
     public function verify(Internship $internship, EntityManagerInterface $entityManager): Response
     {
         $internship->setIsVerified(true);
@@ -100,7 +99,6 @@ final class InternshipController extends AbstractController
 
     #[Route('/{id}', name: 'app_internship_show', methods: ['GET'])]
     #[IsGranted('ROLE_TEACHER')]
-
     public function show(Internship $internship): Response
     {
         return $this->render('internship/show.html.twig', [
@@ -110,9 +108,13 @@ final class InternshipController extends AbstractController
 
     #[Route('/{id}/edit', name: 'app_internship_edit', methods: ['GET', 'POST'])]
     #[IsGranted('ROLE_USER')]
-
     public function edit(Request $request, Internship $internship, EntityManagerInterface $entityManager): Response
     {
+        // Vérifie les permissions avec le voter
+        if (!$this->isGranted('edit', $internship)) {
+            throw new AccessDeniedException('Accès refusé.');
+        }
+
         $form = $this->createForm(InternshipType::class, $internship);
         $form->handleRequest($request);
 
@@ -130,9 +132,12 @@ final class InternshipController extends AbstractController
 
     #[Route('/{id}', name: 'app_internship_delete', methods: ['POST'])]
     #[IsGranted('ROLE_USER')]
-
     public function delete(Request $request, Internship $internship, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('edit', $internship)) {
+            throw new AccessDeniedException('Accès refusé.');
+        }
+
         if ($this->isCsrfTokenValid('delete' . $internship->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($internship);
             $entityManager->flush();
@@ -140,8 +145,4 @@ final class InternshipController extends AbstractController
 
         return $this->redirectToRoute('app_internship_index', [], Response::HTTP_SEE_OTHER);
     }
-
-
-
-
 }
